bullets fire in the direction of where the mouse points, they also face the correct way

+++ Player.h
@@ -6,7 +6,7 @@ public:
     void handleInput(float deltaTime);
     void update(float deltaTime);
     void draw(sf::RenderWindow& window);
-    sf::Vector2i getWeaponPosition();
+    sf::Vector2f getWeaponPosition();
 
 private:
     sf::Sprite sprite;
